requirements.txt
gym-retro
numpy
sounddevice
scipy
python-rtmidi